+++ tests/test_power_spherical.py
@@ -1,6 +1,3 @@
-import sys
-
-import pytest
 import torch
 from power_spherical import HypersphericalUniform, PowerSpherical
 
@@ -51,45 +48,4 @@ def test_kl_divergence():
         (dist1.log_prob(x) - dist2.log_prob(x)).mean(),
         torch.distributions.kl_divergence(dist1, dist2),
         atol=1e-2,
-    )
-
-
-@pytest.mark.xfail(
-    reason="RuntimeError: Encountered autograd state manager op",
-)
-def test_dynamo_export_normal(tmp_path):
-    class Model(torch.nn.Module):
-        def __init__(self):
-            self.normal = torch.distributions.normal.Normal(0, 1)
-            super().__init__()
-
-        def forward(self, x):
-            return self.normal.sample(x.shape)
-
-    x = torch.randn(2, 3)
-    exported_program = torch.export.export(Model(), args=(x,))
-    onnx_program = torch.onnx.dynamo_export(
-        exported_program,
-        x,
-    )
-    onnx_program.save(str(tmp_path / "normal.onnx"))
-
-
-@pytest.mark.xfail(reason="not supported feature of ONNX")
-def test_dynamo_export_spherical():
-    class Model(torch.nn.Module):
-        def __init__(self):
-            self.spherical = PowerSpherical(
-                torch.Tensor([0.0, 1.0]), torch.Tensor([1.0])
-            )
-            super().__init__()
-
-        def forward(self, x):
-            return self.spherical.sample(x.shape)
-
-    x = torch.randn(2, 3)
-    exported_program = torch.export.export(Model(), args=(x,))
-    _ = torch.onnx.dynamo_export(
-        exported_program,
-        x,
     )
